Change ManagedReferenceAttributeDrawer
