+++ Packages/UGF.EditorTools/Editor/IMGUI.References/ManagedReferenceEditorUtility.Deprecated.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 41015120754d4dbab2c3feb656b2dffe
+timeCreated: 1605729983
